// frontend/src/types/generated/TableExportOptions.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { TableExportFormat } from "./TableExportFormat";

export type TableExportOptions = { includeViolationStatus: boolean, includeIds: boolean, omitHeader: boolean, labels: Array<string>, format: TableExportFormat, };